requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bidict==0.23.1
billiard==4.2.0
boto3==1.34.69
botocore==1.34.69
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
daphne==4.1.2
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-filter==24.1
django-magic-link==1.0.0
django-markdownx==4.0.7
django-redis==5.4.0
django-storages==1.14.2
django-templated-mail==1.1.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
dnspython==2.6.1
eventlet==0.36.1
greenlet==3.0.3
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jmespath==1.0.1
kombu==5.3.6
Markdown==3.6
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
pillow==10.2.0
prompt-toolkit==3.0.43
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-http-client==3.3.7
python-socketio==5.11.2
python3-openid==3.2.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.1
sendgrid==6.11.0
sentry-sdk==1.43.0
service-identity==24.1.0
simple-websocket==1.0.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
zope.interface==6.3
